add ContactExists check in GetAllAddressesForUserAsync
This help address a previous bug in the repository wherein fetching all addresses for a contact would fetch all addresses of each of the user's contacts. (instead of just for a specific contact)

A check is added to see if the contact exists. This required a new parameter 'contactId' to be added to the service method. null (not an empty collection) is returned if the contact does not exist.