fix bug in GetEmailByIdForUserAsync in email repository
It was previously okay to request for a valid email id with an invalid contact id.

To fix this, we check if the row's contact's ID matches the given contact ID. This required a new parameter 'contactId' to be added to the service method. 'id' parameter is renamed to 'emailId' for clarity